requirements.txt
ultralytics==8.3.145
torch==2.6.0
torchvision==0.21.0
pyyaml==6.0.2
opencv-python==4.11.0.86
numpy==2.2.3
matplotlib==3.10.1
pandas==2.2.3
Pillow==11.1.0